src/main/resources/python/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0
PyPDF2>=3.0.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0